+++ app/(orgs)/page.tsx
@@ -9,11 +9,13 @@ import { useAppContext } from 'lib/hooks/app-context'
 const IndexPage = () => {
   const router = useRouter()
   const { setActiveBillingOrg } = useAppContext()
-  const [orgs, setOrgs] = useState([])
-  const [isLoadingOrgs, setIsLoadingOrgs] = useState(true)
+  const [ orgs, setOrgs ] = useState([])
+  const [ isLoadingOrgs, setIsLoadingOrgs ] = useState(true)
+  const [ isRedirectingToProjects, setIsRedirectingToProjects ] = useState(false)
 
   useEffect(() => {
     ;(async () => {
+      setIsRedirectingToProjects(false)
       const res = await fetch('/api/orgs')
 
       if (!res.ok) {
@@ -24,6 +26,7 @@ const IndexPage = () => {
         )
       }
       setIsLoadingOrgs(false)
+      setIsRedirectingToProjects(true)
 
       const fetchedOrgs = await res.json()
 
@@ -40,6 +43,7 @@ const IndexPage = () => {
     <div className="flex-1 p-10 text-2xl font-bold">
       <OrgDashboard orgs={orgs} />
       { isLoadingOrgs && <div className="flex justify-center items-center">Loading orgs...</div>}
+      { isRedirectingToProjects && <div className="flex justify-center items-center pt-20">Redirecting to projects...</div>}
     </div>
   )
 }